wip - move primitive numeric types
