+++ packages/overture-schema-base-theme/src/overture/schema/base/infrastructure/models.py
@@ -14,7 +14,7 @@ from overture.schema.core import (
     Feature,
 )
 from overture.schema.core.models import Named, Stacked
-from overture.schema.foundation.primitive.geometry import (
+from overture.schema.foundation.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
