+++ packages/overture-schema-base-theme/src/overture/schema/base/land_use/models.py
@@ -11,7 +11,7 @@ from overture.schema.core import (
     Feature,
 )
 from overture.schema.core.models import Named, Stacked
-from overture.schema.foundation.primitive.geometry import (
+from overture.schema.foundation.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
