+++ packages/overture-schema-buildings-theme/src/overture/schema/buildings/building_part/models.py
@@ -8,7 +8,7 @@ from overture.schema.core import Feature
 from overture.schema.core.models import Named, Stacked
 from overture.schema.core.ref import Reference, Relationship
 from overture.schema.core.types import Id
-from overture.schema.foundation.primitive.geometry import (
+from overture.schema.foundation.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
