+++ packages/overture-schema-buildings-theme/src/overture/schema/buildings/models.py
@@ -8,8 +8,8 @@ from overture.schema.buildings.enums import (
     RoofOrientation,
     RoofShape,
 )
-from overture.schema.core.primitives import float64, int32
 from overture.schema.core.types import HexColor
+from overture.schema.foundation.primitive import float64, int32
 
 
 class Shape(BaseModel):
