+++ packages/overture-schema-core/src/overture/schema/core/models.py
@@ -12,8 +12,10 @@ from pydantic import (
 )
 from pydantic_core import core_schema
 
-from overture.schema.foundation.primitive.bbox import BBox
-from overture.schema.foundation.primitive.geometry import Geometry
+from overture.schema.foundation.primitive import (
+    BBox,
+    Geometry,
+)
 from overture.schema.validation import (
     allow_extension_fields,
 )
