+++ packages/overture-schema-core/tests/test_json_schema_for_primitive_types.py
@@ -1,7 +1,7 @@
 """Tests for JSON Schema generation with primitive types."""
 
 from overture.schema.core.json_schema import json_schema
-from overture.schema.core.primitives import (
+from overture.schema.foundation.primitive import (
     float32,
     float64,
     int32,
