+++ packages/overture-schema-core/tests/test_models.py
@@ -5,8 +5,10 @@ import pytest
 from deepdiff import DeepDiff
 from overture.schema.core.json_schema import EnhancedJsonSchemaGenerator
 from overture.schema.core.models import Feature
-from overture.schema.foundation.primitive.bbox import BBox
-from overture.schema.foundation.primitive.geometry import Geometry
+from overture.schema.foundation.primitive import (
+    BBox,
+    Geometry,
+)
 from shapely.geometry import LineString, Point
