+++ packages/overture-schema-transportation-theme/src/overture/schema/transportation/segment/models.py
@@ -11,7 +11,7 @@ from overture.schema.core.models import (
     Named,
 )
 from overture.schema.core.validation import UniqueItemsConstraint
-from overture.schema.foundation.primitive.geometry import (
+from overture.schema.foundation.primitive import (
     Geometry,
     GeometryType,
     GeometryTypeConstraint,
